strings-ro.xml
Hi. I just translate your english strings.xml in Romanian.
I don't know how to add a new different file with a new name and preserve the existing one - so please if you need to include this translation in your application - do the preparations needed. I propose the name for it to be strings-ro.xml.

Best regards and good luck with your app. Is nice.
Iulian